move IFRT ffi to "IFRT.cpp" file
